+++ src/Ext/Techno/Body.Update.cpp
@@ -986,10 +986,18 @@ void TechnoExt::ExtData::RecalculateStatMultipliers()
 {
 	auto const pThis = this->OwnerObject();
 
-	double firepower = 1.0;
-	double armor = 1.0;
-	double speed = 1.0;
-	double ROF = 1.0;
+	double firepowerMult = 1.0;
+	double armorMult = 1.0;
+	double speedMult = 1.0;
+	double ROFMult = 1.0;
+	int firepowerBonus = 0;
+	int armorBonus = 0;
+	double speedBonus = 0.0;
+	int ROFBonus = 0;
+	int minReceivedDamage = INT32_MIN;
+	int maxReceivedDamage = INT32_MAX;
+	double minSpeed = 0.0;
+	double maxSpeed = INT32_MAX;
 	bool cloak = false;
 	bool forceDecloak = false;
 	bool disableWeapons = false;
@@ -1003,10 +1011,18 @@ void TechnoExt::ExtData::RecalculateStatMultipliers()
 			continue;
 
 		auto const type = attachEffect->GetType();
-		firepower *= type->FirepowerMultiplier;
-		speed *= type->SpeedMultiplier;
-		armor *= type->ArmorMultiplier;
-		ROF *= type->ROFMultiplier;
+		firepowerMult *= type->FirepowerMultiplier;
+		speedMult *= type->SpeedMultiplier;
+		armorMult *= type->ArmorMultiplier;
+		ROFMult *= type->ROFMultiplier;
+		firepowerBonus += type->FirepowerBonus;
+		armorBonus += type->ArmorBonus;
+		speedBonus += type->SpeedBonus;
+		ROFBonus += type->ROFBonus;
+		minReceivedDamage = Math::max(minReceivedDamage, type->ReceivedDamage_Minimum);
+		maxReceivedDamage = Math::min(maxReceivedDamage, type->ReceivedDamage_Maximum);
+		minSpeed = Math::max(minSpeed, type->Speed_Minimum);
+		maxSpeed = Math::min(maxSpeed, type->Speed_Maximum);
 		cloak |= type->Cloakable;
 		forceDecloak |= type->ForceDecloak;
 		disableWeapons |= type->DisableWeapons;
@@ -1015,10 +1031,18 @@ void TechnoExt::ExtData::RecalculateStatMultipliers()
 		reflectsDamage |= type->ReflectDamage;
 	}
 
-	this->AE.FirepowerMultiplier = firepower;
-	this->AE.ArmorMultiplier = armor;
-	this->AE.SpeedMultiplier = speed;
-	this->AE.ROFMultiplier = ROF;
+	this->AE.FirepowerMultiplier = firepowerMult;
+	this->AE.ArmorMultiplier = armorMult;
+	this->AE.SpeedMultiplier = speedMult;
+	this->AE.ROFMultiplier = ROFMult;
+	this->AE.FirepowerBonus = firepowerBonus;
+	this->AE.ArmorBonus = armorBonus;
+	this->AE.SpeedBonus = speedBonus;
+	this->AE.ROFBonus = ROFBonus;
+	this->AE.ReceivedDamage_Minimum = minReceivedDamage;
+	this->AE.ReceivedDamage_Maximum = maxReceivedDamage;
+	this->AE.Speed_Minimum = minSpeed;
+	this->AE.Speed_Maximum = maxSpeed;
 	this->AE.Cloakable = cloak;
 	this->AE.ForceDecloak = forceDecloak;
 	this->AE.DisableWeapons = disableWeapons;